+++ TileEngine/Tile Animation.cpp
@@ -557,7 +557,7 @@ void UpdateAniTiles( )
 					// CHECK IF WE SHOULD BE DISPLAYING TRANSLUCENTLY!
 					if ( pNode->sCurrentFrame == pNode->ubKeyFrame2 )
 					{
-						UINT8     ubExpType;
+						UINT16     ubExpType;
 
 						switch( pNode->uiKeyFrame2Code )
 						{
@@ -576,7 +576,7 @@ void UpdateAniTiles( )
 							}
 							else
 							{
-								SpreadEffect( pNode->sGridNo, Explosive[ Item[ (UINT16)pNode->uiUserData ].ubClassIndex ].ubRadius, (UINT16)pNode->uiUserData, (UINT8)pNode->ubUserData2, FALSE, gExplosionData[ pNode->uiUserData3 ].Params.bLevel, -1 );
+								SpreadEffect( pNode->sGridNo, (UINT8)Explosive[ Item[ (UINT16)pNode->uiUserData ].ubClassIndex ].ubRadius, (UINT16)pNode->uiUserData, (UINT8)pNode->ubUserData2, FALSE, gExplosionData[ pNode->uiUserData3 ].Params.bLevel, -1 );
 							}
 							// Forfait any other animations this frame....
 							return;
